# docs/openapi.yml
openapi: 3.1.0
info:
  title: Aegiscan API
  summary: Aegiscan API
  description: Aegiscan is the open source Tines / Splunk SOAR alternative. You can
    operate Aegiscan in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Aegiscan Founders
    email: founders@digi-trans.org
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /actions:
    get:
      tags:
      - actions
      summary: List Actions
      description: List all actions for a workflow.
      operationId: actions-list_actions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: workflow_id
        in: query
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionReadMinimal'
                title: Response Actions-List Actions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - actions
      summary: Create Action
      description: Create a new action for a workflow.
      operationId: actions-create_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionReadMinimal'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /actions/{action_id}:
    get:
      tags:
      - actions
      summary: Get Action
      description: Get an action.
      operationId: actions-get_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          pattern: act-[0-9a-f]{32}
          title: Action Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: workflow_id
        in: query
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - actions
      summary: Update Action
      description: Update an action.
      operationId: actions-update_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          pattern: act-[0-9a-f]{32}
          title: Action Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - actions
      summary: Delete Action
      description: Delete an action.
      operationId: actions-delete_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          pattern: act-[0-9a-f]{32}
          title: Action Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /organization/members:
    get:
      tags:
      - organization
      summary: List Org Members
      operationId: organization-list_org_members
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrgMemberRead'
                type: array
                title: Response Organization-List Org Members
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /organization/members/{user_id}:
    delete:
      tags:
      - organization
      summary: Delete Org Member
      operationId: organization-delete_org_member
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - organization
      summary: Update Org Member
      operationId: organization-update_org_member
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMemberRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /organization/sessions:
    get:
      tags:
      - organization
      summary: List Sessions
      operationId: organization-list_sessions
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SessionRead'
                type: array
                title: Response Organization-List Sessions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /organization/sessions/{session_id}:
    delete:
      tags:
      - organization
      summary: Delete Session
      operationId: organization-delete_session
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Session Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /editor/actions:
    get:
      tags:
      - editor
      summary: List Actions
      operationId: editor-list_actions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: workflow_id
        in: query
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditorActionRead'
                title: Response Editor-List Actions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /registry/actions:
    get:
      tags:
      - registry-actions
      summary: List Registry Actions
      description: List all actions in a registry.
      operationId: registry-actions-list_registry_actions
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RegistryActionReadMinimal'
                type: array
                title: Response Registry-Actions-List Registry Actions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    post:
      tags:
      - registry-actions
      summary: Create Registry Action
      description: Create a new registry action.
      operationId: registry-actions-create_registry_action
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistryActionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryActionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /registry/actions/{action_name}:
    get:
      tags:
      - registry-actions
      summary: Get Registry Action
      description: Get a specific registry action.
      operationId: registry-actions-get_registry_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_name
        in: path
        required: true
        schema:
          type: string
          title: Action Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryActionRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - registry-actions
      summary: Update Registry Action
      description: Update a custom registry action.
      operationId: registry-actions-update_registry_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_name
        in: path
        required: true
        schema:
          type: string
          title: Action Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistryActionUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - registry-actions
      summary: Delete Registry Action
      description: Delete a template action.
      operationId: registry-actions-delete_registry_action
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: action_name
        in: path
        required: true
        schema:
          type: string
          title: Action Name
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ActionControlFlow:
      properties:
        run_if:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Run If
        for_each:
          anyOf:
          - anyOf:
            - type: string
            - items:
                type: string
              type: array
            maxLength: 1000
          - type: 'null'
          title: For Each
        join_strategy:
          allOf:
          - $ref: '#/components/schemas/JoinStrategy'
          default: all
        retry_policy:
          $ref: '#/components/schemas/ActionRetryPolicy'
        start_delay:
          type: number
          title: Start Delay
          description: Delay before starting the action in seconds. If `wait_until`
            is also provided, the `wait_until` timer will take precedence.
          default: 0.0
        wait_until:
          anyOf:
          - type: string
          - type: 'null'
          title: Wait Until
          description: Wait until a specific date and time before starting. Overrides
            `start_delay` if both are provided.
      type: object
      title: ActionControlFlow
    ActionCreate:
      properties:
        workflow_id:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
        type:
          type: string
          title: Type
        title:
          type: string
          maxLength: 100
          minLength: 1
          title: Title
      type: object
      required:
      - workflow_id
      - type
      - title
      title: ActionCreate
    ActionRead:
      properties:
        id:
          type: string
          pattern: act-[0-9a-f]{32}
          title: Id
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        inputs:
          type: string
          title: Inputs
        control_flow:
          $ref: '#/components/schemas/ActionControlFlow'
      type: object
      required:
      - id
      - type
      - title
      - description
      - status
      - inputs
      title: ActionRead
    ActionReadMinimal:
      properties:
        id:
          type: string
          pattern: act-[0-9a-f]{32}
          title: Id
        workflow_id:
          type: string
          pattern: wf_[0-9a-zA-Z]+
          title: Workflow Id
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
      type: object
      required:
      - id
      - workflow_id
      - type
      - title
      - description
      - status
      title: ActionReadMinimal
...
    ActionUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        inputs:
          type: string
          maxLength: 10000
          title: Inputs
          default: ''
        control_flow:
          anyOf:
          - $ref: '#/components/schemas/ActionControlFlow'
          - type: 'null'
      type: object
      title: ActionUpdate
...
    EditorActionRead:
      properties:
        type:
          type: string
          title: Type
        ref:
          type: string
          title: Ref
        description:
          type: string
          title: Description
      type: object
      required:
      - type
      - ref
      - description
      title: EditorActionRead
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JoinStrategy:
      type: string
      enum:
      - any
      - all
      title: JoinStrategy
...
    OrgMemberRead:
      properties:
        user_id:
          type: string
          format: uuid4
          title: User Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        is_verified:
          type: boolean
          title: Is Verified
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
      type: object
      required:
      - user_id
      - first_name
      - last_name
      - email
      - role
      - is_active
      - is_superuser
      - is_verified
      - last_login_at
      title: OrgMemberRead
    RegistryActionCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: The name of the action
        description:
          type: string
          maxLength: 1000
          title: Description
          description: The description of the action
        namespace:
          type: string
          title: Namespace
          description: The namespace of the action
        type:
          type: string
          enum:
          - udf
          - template
          title: Type
          description: The type of the action
        origin:
          type: string
          maxLength: 1000
          minLength: 1
          title: Origin
          description: The origin of the action as a url
        secrets:
          anyOf:
          - items:
              $ref: '#/components/schemas/RegistrySecret'
            type: array
          - type: 'null'
          title: Secrets
          description: The secrets required by the action
        interface:
          $ref: '#/components/schemas/RegistryActionInterface'
        implementation:
          anyOf:
          - $ref: '#/components/schemas/RegistryActionTemplateImpl-Input'
          - $ref: '#/components/schemas/RegistryActionUDFImpl'
          title: Implementation
        default_title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Default Title
          description: The default title of the action
        display_group:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Display Group
          description: The presentation group of the action
        doc_url:
          anyOf:
          - type: string
            maxLength: 1000
            minLength: 1
          - type: 'null'
          title: Doc Url
          description: Link to documentation
        author:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Author
          description: Author of the action
        deprecated:
          anyOf:
          - type: string
            maxLength: 1000
            minLength: 1
          - type: 'null'
          title: Deprecated
          description: Marks action as deprecated along with message
        options:
          allOf:
          - $ref: '#/components/schemas/RegistryActionOptions'
          description: The options for the action
        repository_id:
          type: string
          format: uuid4
          title: Repository Id
          description: The repository id
      type: object
      required:
      - name
      - description
      - namespace
      - type
      - origin
      - interface
      - repository_id
      title: RegistryActionCreate
      description: API create model for a registered action.
...
    RegistryActionReadMinimal:
      properties:
        name:
          type: string
          title: Name
          description: The name of the action
        description:
          type: string
          title: Description
          description: The description of the action
        namespace:
          type: string
          title: Namespace
          description: The namespace of the action
        type:
          type: string
          enum:
          - udf
          - template
          title: Type
          description: The type of the action
        origin:
          type: string
          title: Origin
          description: The origin of the action as a url
        default_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Title
          description: The default title of the action
        display_group:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Group
          description: The presentation group of the action
        action:
          type: string
          title: Action
          description: The full action identifier.
          readOnly: true
      type: object
      required:
      - name
      - description
      - namespace
      - type
      - origin
      - action
      title: RegistryActionReadMinimal
      description: API minimal read model for a registered action.
...
    SessionRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        user_id:
          type: string
          format: uuid4
          title: User Id
        user_email:
          type: string
          format: email
          title: User Email
      type: object
      required:
      - id
      - created_at
      - user_id
      - user_email
      title: SessionRead
...
    UserRole:
      type: string
      enum:
      - basic
      - admin
      title: UserRole
    UserUpdate:
      properties:
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_superuser:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Superuser
        is_verified:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Verified
        role:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        settings:
          anyOf:
          - type: object
          - type: 'null'
          title: Settings
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-aegiscan-service-key
...